refactor: remove legacy AI Platform Engineering charts and introduce a composite Vault chart
- Deleted the old AI Platform Engineering charts (v1 and v2) to streamline the codebase.
- Introduced a new Vault composite chart that encapsulates all Vault-related components, enhancing modularity and maintainability.
- Updated main entry point to reflect the removal of old charts and the addition of the new Vault chart.
